# Performance: remove stack safety from Http.execute