--- a/zio-http-benchmarks/src/main/scala/zhttp.benchmarks/HttpCollectEval.scala
+++ b/zio-http-benchmarks/src/main/scala/zhttp.benchmarks/HttpCollectEval.scala
@@ -15,7 +15,7 @@ class HttpCollectEval {
 
   @Benchmark
   def benchmarkApp(): Unit = {
-    (0 to MAX).foreach(_ => app.execute(0).evaluate)
+    (0 to MAX).foreach(_ => app.execute(0))
     ()
   }