--- a/zio-http-benchmarks/src/main/scala/zhttp.benchmarks/HttpCombineEval.scala
+++ b/zio-http-benchmarks/src/main/scala/zhttp.benchmarks/HttpCombineEval.scala
@@ -15,13 +15,13 @@ class HttpCombineEval {
 
   @Benchmark
   def benchmarkNotFound(): Unit = {
-    spec.execute(-1).evaluate
+    spec.execute(-1)
     ()
   }
 
   @Benchmark
   def benchmarkOk(): Unit = {
-    spec.execute(0).evaluate
+    spec.execute(0)
     ()
   }
 }