--- a/zio-http-benchmarks/src/main/scala/zhttp.benchmarks/HttpNestedFlatMapEval.scala
+++ b/zio-http-benchmarks/src/main/scala/zhttp.benchmarks/HttpNestedFlatMapEval.scala
@@ -10,7 +10,7 @@ import java.util.concurrent.TimeUnit
 @OutputTimeUnit(TimeUnit.SECONDS)
 class HttpNestedFlatMapEval {
 
-  private val MAX = 10000
+  private val MAX = 1000
 
   val programFlatMap: Http[Any, Nothing, Int, Int] =
     (0 to MAX).foldLeft(Http.identity[Int])((a, _) => a.flatMap(i => Http.succeed(i + 1)))