--- a/zio-http-test/src/main/scala/zhttp/test/test.scala
+++ b/zio-http-test/src/main/scala/zhttp/test/test.scala
@@ -5,6 +5,6 @@ import zio.ZIO
 
 package object test {
   implicit class HttpWithTest[R, E, A, B](http: Http[R, E, A, B]) {
-    def apply(req: A): ZIO[R, Option[E], B] = http.execute(req).evaluate.asEffect
+    def apply(req: A): ZIO[R, Option[E], B] = http.execute(req).toEffect
   }
 }