There are two optimizations done here — 
1. Operators on `HExit` now are using final encoding. This reduces allocations and improves performances dramatically in certain use-cases.
2. Remove stack safety from `Http`. This is done because there are no "real" use-cases where you would need to recursively create `Http`. Moreover adding stack safety by using the heap, slows down the evaluation of the domain.

# Benchmarks

**NOTE:**

- Iteration count was reduced to `1000` from `10,000` because we are losing out on stack safety.

**Before**
```
[info] Benchmark                                         Mode   Cnt       Score       Error  Units
[info] HttpNestedFlatMapEval.benchmarkHttpFlatMap        thrpt    3   12335.029 ± 13513.481  ops/s
```

**After**
```
[info] Benchmark                                         Mode   Cnt        Score      Error  Units
[info] HttpNestedFlatMapEval.benchmarkHttpFlatMap        thrpt    3   113165.055 ± 9829.742  ops/s
```



**Before**
```
[info] Benchmark                           Mode  Cnt      Score       Error  Units
[info] HttpCombineEval.benchmarkNotFound  thrpt    3  16582.192 ±  5952.810  ops/s
[info] HttpCombineEval.benchmarkOk        thrpt    3  13392.093 ± 35982.678  ops/s
```

**After**
```
[info] Benchmark                           Mode  Cnt       Score       Error  Units
[info] HttpCombineEval.benchmarkNotFound  thrpt    3  121755.257 ± 56679.664  ops/s
[info] HttpCombineEval.benchmarkOk        thrpt    3   49160.038 ± 39910.239  ops/s
````