@amitksingh1490 2021-12-13
## 🚀🚀🚀 Benchmark Results 
 **Concurrency: 256** 
 **Requests/sec: 440238.36**