@amitksingh1490 2021-12-14
## 🚀🚀🚀 Benchmark Results 
 **Concurrency: 256** 
 **Requests/sec: 434724.49**