@amitksingh1490 2021-12-15
## 🚀🚀🚀 Benchmark Results 
 **Concurrency: 256** 
 **Requests/sec: 443408.09**