@amitksingh1490 2021-12-16
## 🚀🚀🚀 Benchmark Results 
 **Concurrency: 256** 
 **Requests/sec: 444126.41**